api/requirements.txt
azure-functions
psycopg[binary,pool]
python-dotenv
groq
google-generativeai
pyjwt
cryptography
requests
psycopg2-binary
redis
azure-cognitiveservices-speech
azure-storage-blob
gTTS
svix
